+++ src/extremeDevice.java
@@ -1,6 +1,6 @@
 public class extremeDevice extends Device {
 
-    public extremeDevice(float minWeight, float minHeight, int minAge, boolean isOpen, String name, float price) {
+    public extremeDevice(double minWeight, double minHeight, int minAge, boolean isOpen, String name, double price) {
         super(minWeight, minHeight, minAge, isOpen, name, price);
     }
 }
